Use extracted functions instead of repeating ourselves
